requirements.txt
# 核心依赖
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0

# 数据库
PyMySQL==1.1.0
cryptography==42.0.0

# 生产环境
waitress==2.1.2
gunicorn==21.2.0

# 开发测试
pytest==7.4.4
pytest-flask==1.3.0
black==24.1.1
flake8==7.0.0
